Move RangeNode and RangeTree logic to core
It contains no range-specific code and could ideally be used from core.
