+++ src/nc/core/RangeTree.cpp
@@ -4,7 +4,7 @@
 #include "RangeTree.h"
 
 namespace nc {
-namespace gui {
+namespace core {
 
 RangeTree::RangeTree() {}
 
@@ -163,6 +163,6 @@ std::vector<const RangeNode *> RangeTree::handleInsertion(int position, int ncha
     return result;
 }
 
-}} // namespace nc::gui
+}} // namespace nc::core
 
 /* vim:set et sts=4 sw=4: */
