+++ src/nc/core/RangeTree.h
@@ -15,9 +15,7 @@ namespace nc {
 namespace core { namespace likec {
     class TreeNode;
     class Tree;
-}}
-
-namespace gui {
+}
 
 class RangeTree {
     std::unique_ptr<RangeNode> root_;
@@ -38,6 +36,6 @@ public:
     std::vector<const RangeNode *> handleInsertion(int position, int nchars);
 };
 
-}} // namespace nc::gui
+}} // namespace nc::core
 
 /* vim:set et sts=4 sw=4: */
